#23 setting: data 모듈 폴더구조 변경
