+++ data/src/main/java/com/exhibitiondot/data/datasource/event/EventDataSource.kt
@@ -1,4 +1,4 @@
-package com.exhibitiondot.data.datasource
+package com.exhibitiondot.data.datasource.event
 
 import androidx.paging.PagingData
 import com.exhibitiondot.data.model.dto.EventDetailDto
